Модулям даны более подходящие имена
